# src/main/resources/api/openapi.yml
openapi: 3.1.0
info:
  title: message-exchange
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.3"
servers:
- url: http://localhost:45109
  description: Generated server url
...
paths:
  /{municipalityId}/{namespace}/conversations:
    get:
      tags:
      - Conversation resource
      description: Get all conversations with or without filters. The resource allows
        the client a wide range of variations on how to filter the result.
      operationId: getConversations
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: MY_NAMESPACE
      - name: filter
        in: query
        description: "Syntax description: [spring-filter](https://github.com/turkraft/spring-filter/blob/85730f950a5f8623159cc0eb4d737555f9382bb7/README.md#syntax)"
        required: false
        schema:
          type: string
        example: topic:'My topic' and messages.createdBy.value:'joe01doe' and messages.created>'2023-01-01T00:00:00Z'
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageConversation"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Conversation resource
      description: Create a conversation
      operationId: createConversation
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: MY_NAMESPACE
      requestBody:
        description: Conversation to be created
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Conversation"
        required: true
      responses:
        "201":
          description: Created - Successful operation
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/conversations/{conversationId}:
    get:
      tags:
      - Conversation resource
      description: Get a conversation
      operationId: getConversation
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: MY_NAMESPACE
      - name: conversationId
        in: path
        description: Conversation ID
        required: true
        schema:
          type: string
        example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Conversation resource
      description: Delete a conversation
      operationId: deleteConversation
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: MY_NAMESPACE
      - name: conversationId
        in: path
        description: Conversation ID
        required: true
        schema:
          type: string
        example: b82bd8ac-1507-4d9a-958d-369261eecc15
      responses:
        "204":
          description: No Content - Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Conversation resource
      description: Update a conversation
      operationId: updateConversation
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: MY_NAMESPACE
      - name: conversationId
        in: path
        description: Conversation ID
        required: true
        schema:
          type: string
        example: b82bd8ac-1507-4d9a-958d-369261eecc15
      requestBody:
        description: Conversation to be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Conversation"
        required: true
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{namespace}/conversations/{conversationId}/countReadBy:
    get:
      tags:
      - Conversation resource
      description: "Get read statistics for a conversation, describing how many messages\
        \ have been read per identifier and per part."
      operationId: countReadBy
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: namespace
        in: path
        description: Namespace
        required: true
        schema:
          type: string
        example: MY_NAMESPACE
      - name: conversationId
        in: path
        description: Conversation ID
        required: true
        schema:
          type: string
        example: b82bd8ac-1507-4d9a-958d-369261eecc15
      - name: includeSystemMessages
        in: query
        description: Whether system generated messages should be included in the statistics.
          Defaults to false.
        required: false
        schema:
          type: boolean
          default: false
        example: false
      responses:
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadByStatistics"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Conversation:
      type: object
      description: "Represents a conversation containing identifiers, messages, and\
        \ metadata."
      properties:
        id:
          type: string
          description: The unique identifier of the conversation.
          examples:
          - 2af1002e-008f-4bdc-924b-daaae31f1118
          readOnly: true
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Identifier"
            description: The list of participants in the conversation.
        municipalityId:
          type: string
          description: The identifier of the municipality associated with the conversation.
          examples:
          - "2281"
          readOnly: true
        namespace:
          type: string
          description: The namespace of the conversation.
          examples:
          - namespace-abc
          readOnly: true
        externalReferences:
          type: array
          items:
            $ref: "#/components/schemas/KeyValues"
            description: The list of external references associated with the conversation.
        metadata:
          type: array
          items:
            $ref: "#/components/schemas/KeyValues"
            description: A list of metadata associated with the conversation.
        topic:
          type: string
          description: The topic of the conversation.
          examples:
          - Customer Support
        latestSequenceNumber:
          type: integer
          format: int64
          description: The latest sequence number of the conversation.
          examples:
          - 1
          readOnly: true
    Identifier:
      type: object
      description: "Represents a participant in a conversation, including their type\
        \ and value."
      properties:
        type:
          type: string
          description: "The type of the participant, e.g., adAccount or partyId"
          examples:
          - user
          pattern: ^(adAccount|partyId)$
        value:
          type: string
          description: The unique identifier or value of the participant.
          examples:
          - ad012ad
          minLength: 1
      required:
      - value
    KeyValues:
      type: object
      description: Represents a key-value pair of metadata.
      properties:
        key:
          type: string
          description: The key of the data
          examples:
          - key1
        values:
          type: array
          items:
            type: string
            description: The values associated with the key
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        detail:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        instance:
          type: string
          format: uri
        detail:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    PageConversation:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
        number:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        first:
          type: boolean
        last:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    ReadByCount:
      type: object
      description: The number of messages read by a specific identifier within a conversation.
      properties:
        identifier:
          $ref: "#/components/schemas/Identifier"
          description: The identifier of the person the count refers to.
        count:
          type: integer
          format: int64
          description: The number of messages read by the identifier.
          examples:
          - 4
    ReadByPartCount:
      type: object
      description: The number of messages read by a specific part within a conversation.
      properties:
        part:
          type: string
          description: The part the count refers to.
          examples:
          - errand-123
        count:
          type: integer
          format: int64
          description: The number of messages read by the part.
          examples:
          - 6
    ReadByStatistics:
      type: object
      description: "Read statistics for a conversation, describing how many messages\
        \ have been read per identifier and per part."
      properties:
        messageCount:
          type: integer
          format: int64
          description: The total number of messages the statistics are based on.
          examples:
          - 13
          readOnly: true
        readByCount:
          type: array
          items:
            $ref: "#/components/schemas/ReadByCount"
            description: The number of messages read per identifier.
        readByPartCount:
          type: array
          items:
            $ref: "#/components/schemas/ReadByPartCount"
            description: The number of messages read per part.
  securitySchemes: {}